refactor(ui): move tierOptions into complexity_router_tiers, the shared tier-utility module
